indexer/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy
